Server/requirements/production.txt
-r base.txt


# Database & Caching
psycopg2-binary==2.9.10  # PostgreSQL adapter
redis==5.2.1              # For session caching and rate limiting
django-redis==5.4.0       # Redis cache backend for Django


# Performance Monitoring
django-silk==5.3.0         # Django profiling and performance monitoring
prometheus-client==0.24.1  # Metrics collection
django-prometheus==2.4.0   # Prometheus integration for Django